Update views.py
Changes were made to avoid having the paginator function in every method. Also we resolved the issue with the urls having the "/update_complete" for example, by redirecting to "db" after every insert or update. Now we have to issue to change the page (with Next>>) in "Reports" tab, because of the wrong url.